+++ tools/scripts/format.ts
@@ -1,4 +1,4 @@
-import { RunSync } from '../../src/Platform/Bun/Child Process.js';
+import { RunSync } from '../../src/lib/ericchase/Platform/Bun/Child Process.js';
 import { command_map } from '../dev.js';
 import { Cache_FileStats_Lock, Cache_FileStats_Unlock } from '../lib/cache/FileStatsCache.js';
 import { TryLock } from '../lib/cache/LockCache.js';